rotation averaging unittest

+++ rotation_averaging/rotation_averaging.h
@@ -12,7 +12,7 @@
 class RotationAveraging {
 public:
   static Eigen::Quaterniond averageRotations(
-      const std::map<std::pair<int, int>, Eigen::Quaterniond> &relative_rotations,
+      const std::map<std::pair<size_t, size_t>, Eigen::Quaterniond> &relative_rotations,
       size_t num_of_rotations, std::vector<Eigen::Quaterniond> *rotations);
 };
